% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Show the location of fatal accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A state number}

\item{year}{A four digit year}
}
\description{
Given a state number and a four digit year, a map is generated with
one data marker at the location of each fatal accident. If the state number is
invalid, or if no records are found for that state, an error is returned
}
\note{
If there are multiple accidents at a location, there will be only
    one data point at that location on the map
}
\examples{
    \dontrun{
  fars_map_state(1,2014)
  fars_map_state("3",2012)
  fars_map_state(3, "2012")
}
}
