% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Generate tibbles to count fatal accidents by month and year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of four-digit years}
}
\value{
A list of tibbles, one list element for each year in years. Each tibble has one row per fatal accident,
         containing the month and year of the accident
}
\description{
For each year in the input list, reads the fars data file for that year, and returns a
tibble with one row per fatal accident. The only data in the row is the year and month of the
accident.  The tibbles are only useful for later counting by month and year.
Returns a NULL list element if a year in the list is not found in the local data,
and issues a warning.
}
\examples{
   fars_read_years(list(2012))
   fars_read_years(list("2012"))
   fars_read_years(list(2012:2014))

}
