% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Create a tibble with number of fatal accidents per month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of four-digit years}
}
\value{
tibble with number of fatal accidents in a tabular format
}
\description{
The input is a list of four digit years. The output is a tibble with one row per month,
and one column per year of the number of fatal accidents. If any of the input values
for year cannot be interpreted by \code{as.integer}, or if there is not a data
file for a year, a message is printed warning of an invalid year. Valid years are
processed
}
\examples{
 fars_summarize_years(list(2012))
 fars_summarize_years(list("2012","2014"))
 fars_summarize_years(list(2012:2014))
}
