% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a fars data file name from a year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The four-digit year for which data is desired}
}
\value{
character string of fars file name for given year
}
\description{
Given a year, returns a string in the format of a fars data file
name for that year
}
\note{
Does not check for valid year or existence of data file
}
\examples{
 make_filename(2013)
 make_filename("2013")

}
